personal_data/requirements.txt
bcrypt==4.0.1
mysql-connector-python==8.0.33
flask==2.3.2
pytest==7.3.1
